Fixed bug in pulling out solitary x

+++ Mariasek.Engine.New/Probability.cs
@@ -409,7 +409,7 @@ namespace Mariasek.Engine.New
             //pokud mam jiste nejakou z karet na seznamu, potom je pravdepodobnost rovna pravdepodobnosti, ze nemam zadnou z karet co neni na seznamu
             if (listedCertainCardsInSuit > 0)
             {
-                return CNK(uncertainCardsNotInSuit, unknownCards) / CNK(uncertainCards, unknownCards);
+                return CNK(uncertainCardsNotInSuit, unknownCards) / (float)CNK(uncertainCards, unknownCards);
             }
 
             //do citatele dame vsechny k-tice ktere hrac v dane barve muze mit a do jmenovatele vsechny kombinace
